requirements.txt
keyboard==0.13.5
PyYAML==6.0.1
selenium==4.16.0
